requirements.txt
mysql-connector-python>=8.0.0
pandas>=1.5.0
tqdm>=4.0.0
python-dotenv>=0.20.0